app/requirements.txt
pandas
numpy
requests
python-dateutil
boto3